+++ CHANGELOG.md
@@ -25,6 +25,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
   - `progress`;
   - `remove`;
   - `update`;
+- Updated `add` command to set 'createdAt' and 'updatedAt' properties.
+- Updated `complete` command to set 'updatedAt' property.
+- Updated `update` command to set 'updatedAt' property.
 - Updated return values for the following commands:
   - `backup`;
   - `update`;